+++ Tests/KeyboardKitTests/Feedback/FeedbackSettingsTests.swift
@@ -23,8 +23,8 @@ class FeedbackSettingsTests: QuickSpec {
             }
             
             it("can use custom configurations") {
-                let audio = AudioFeedbackConfiguration(inputFeedback: .delete, deleteFeedback: .input, systemFeedback: .system)
-                let haptic = HapticFeedbackConfiguration(tapFeedback: .error, doubleTapFeedback: .warning, longPressFeedback: .success, longPressOnSpaceFeedback: .lightImpact, repeatFeedback: .error)
+                let audio = AudioFeedbackConfiguration(input: .delete, delete: .input, system: .system)
+                let haptic = HapticFeedbackConfiguration(tap: .error, doubleTap: .warning, longPress: .success, longPressOnSpace: .lightImpact, repeat: .error)
                 let settings = KeyboardFeedbackSettings(audioConfiguration: audio, hapticConfiguration: haptic)
                 expect(settings.audioConfiguration).toNot(equal(.standard))
                 expect(settings.hapticConfiguration).toNot(equal(.standard))